+++ src/fields/SeoField.php
@@ -148,9 +148,8 @@ class SeoField extends Field implements PreviewableFieldInterface
 
 		// Filter out empty robots
 		if (array_key_exists('robots', $value['advanced']))
-		{
-			$value['advanced']['robots'] = array_filter($value['advanced']['robots']);
-		}
+			$value['advanced']['robots'] =
+				array_filter($value['advanced']['robots']);
 
 		return $value;
 	}